+++ src/composeApp/build.gradle.kts
@@ -24,10 +24,8 @@ kotlin {
             implementation(projects.platform)
             implementation(projects.featureWishlist)
             implementation(projects.featureSearch)
-            implementation(projects.featureDetails)
             implementation(projects.featureMovies)
             implementation(projects.featureTvshows)
-            implementation(projects.featureTvshowdetails)
             implementation(projects.data)
             implementation(projects.domain)
         }
