--- src/feature-details/build.gradle.kts
@@ -1,30 +1,0 @@
-plugins {
-    id("kmp-library-plugin")
-    alias(libs.plugins.jetbrains.compose)
-    alias(libs.plugins.compose.compiler)
-}
-
-kotlin {
-    sourceSets {
-        commonMain.dependencies {
-            implementation(libs.bundles.koin)
-
-            implementation(libs.bundles.compose.multiplatform)
-
-            implementation(libs.kotlinx.coroutines.core)
-            implementation(libs.lifecycle.viewmodel.compose)
-            implementation(libs.kotlinx.collections.immutable)
-            implementation(libs.navigation.compose)
-
-            implementation(projects.domain)
-            implementation(projects.designsystem)
-            implementation(projects.platform)
-        }
-
-        commonTest.dependencies {
-            implementation(libs.bundles.test.multiplatform)
-            implementation(libs.kotlinx.coroutines.core)
-            implementation(libs.kotlin.stdlib)
-        }
-    }
-}
