--- src/feature-details/src/commonMain/composeResources/values/strings.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- Details -->
-    <string name="alt_is_favorite">It is Favorite</string>
-    <string name="alt_is_not_favorite">It is not favorite</string>
-    <string name="overview">Overview</string>
-    <string name="popularity">Popularity</string>
-    <string name="language">Language</string>
-    <string name="tagline">Tagline</string>
-    <string name="production_companies">Production companies</string>
-    <string name="homepage">Homepage</string>
-    <string name="poster">Poster</string>
-</resources>
