+++ src/feature-movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/feature/movies/di/MoviesModule.kt
@@ -1,6 +1,7 @@
 package com.gabrielbmoro.moviedb.feature.movies.di
 
 import MoviesHandler
+import com.gabrielbmoro.moviedb.feature.movies.ui.screens.details.DetailsViewModel
 import com.gabrielbmoro.moviedb.feature.movies.ui.screens.movies.MoviesViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
@@ -18,6 +19,14 @@ val featureMoviesModule = lazyModule {
             ioDispatcher = Dispatchers.IO,
             loggerHelper = get(),
             moviesHandler = get(),
+        )
+    }
+    viewModel<DetailsViewModel> {
+        DetailsViewModel(
+            getMovieDetailsUseCase = get(),
+            favoriteMovieUseCase = get(),
+            repository = get(),
+            ioDispatcher = Dispatchers.IO,
         )
     }
 }
