+++ src/feature-movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/feature/movies/ui/screens/details/Model.kt
@@ -1,0 +1,47 @@
+package com.gabrielbmoro.moviedb.feature.movies.ui.screens.details
+
+import com.gabrielbmoro.moviedb.platform.viewmodel.UiState
+import com.gabrielbmoro.moviedb.platform.viewmodel.UserIntent
+import kotlinx.collections.immutable.ImmutableList
+
+sealed interface DetailsUserIntent : UserIntent {
+    data object HideVideo : DetailsUserIntent
+
+    data object FavoriteMovie : DetailsUserIntent
+
+    data class LoadMovieDetails(
+        val movieId: Long,
+    ) : DetailsUserIntent
+}
+
+data class DetailsUIState(
+    val movieTitle: String,
+    val isLoading: Boolean = true,
+    val isFavorite: Boolean,
+    val movieVotesAverage: Float,
+    val movieLanguage: String,
+    val moviePopularity: Float,
+    val movieOverview: String,
+    val imageUrl: String?,
+    val tagLine: String? = null,
+    val genres: ImmutableList<String>? = null,
+    val status: String? = null,
+    val productionCompanies: String? = null,
+    val homepage: String? = null,
+    val videoId: String? = null,
+    val showVideo: Boolean = true,
+    val errorMessage: String? = null,
+) : UiState {
+    companion object {
+        fun empty() =
+            DetailsUIState(
+                isFavorite = false,
+                movieTitle = "",
+                moviePopularity = 0f,
+                movieLanguage = "",
+                movieVotesAverage = 0f,
+                movieOverview = "",
+                imageUrl = "",
+            )
+    }
+}
