--- src/feature-tvshowdetails/src/commonMain/composeResources/values/strings.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="overview">Overview</string>
-    <string name="popularity">Popularity</string>
-    <string name="language">Language</string>
-    <string name="tagline">Tagline</string>
-    <string name="first_air_date">First Air Date</string>
-    <string name="last_air_date">Last Air Date</string>
-    <string name="status">Status</string>
-    <string name="seasons">Seasons</string>
-    <string name="episodes">Episodes</string>
-    <string name="networks">Networks</string>
-    <string name="created_by">Created By</string>
-    <string name="production_companies">Production companies</string>
-    <string name="homepage">Homepage</string>
-    <string name="poster">Poster</string>
-    <string name="retry">Retry</string>
-</resources>
