--- src/feature-tvshowdetails/src/commonMain/kotlin/com/gabrielbmoro/moviedb/feature/tvshowdetails/di/TvShowDetailsModule.kt
@@ -1,17 +1,0 @@
-package com.gabrielbmoro.moviedb.feature.tvshowdetails.di
-
-import com.gabrielbmoro.moviedb.feature.tvshowdetails.ui.screens.details.TvShowDetailsViewModel
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
-import org.koin.core.module.dsl.viewModel
-import org.koin.dsl.lazyModule
-
-val featureTvShowDetailsModule = lazyModule {
-    viewModel<TvShowDetailsViewModel> {
-        TvShowDetailsViewModel(
-            getTvShowDetailsUseCase = get(),
-            ioDispatcher = Dispatchers.IO,
-            loggerHelper = get(),
-        )
-    }
-}
